% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example2_data.R
\docType{data}
\name{example2_data}
\alias{example2_data}
\title{Mackey-Glass with added noise}
\format{
An object of class \code{numeric} of length 4999.
}
\usage{
example2_data
}
\description{
4999 data points
}
\examples{
ex2_embedded <- embed(as.matrix(example2_data), 25)
targets <- ex2_embedded[ ,1]
predictors <- ex2_embedded[ ,2:25]
gamma_test(predictors, targets)
}
\references{
Described in vignette 1_time_series_with_sr
}
\keyword{datasets}
