% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgls.R
\docType{data}
\name{mgls}
\alias{mgls}
\title{Mackey-Glass time delayed differential equation}
\format{
An object of class \code{numeric} of length 4999.
}
\usage{
mgls
}
\description{
4999 data points
}
\examples{
mgls_embedded <- embed(as.matrix(mgls), 25)
targets <- mgls_embedded[ ,1]
predictors <- mgls_embedded[ ,2:25]
}
\references{
See Wikipedia entry on "Mackey-Glass equations"
}
\keyword{datasets}
