% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_by_mask.R
\name{select_by_mask}
\alias{select_by_mask}
\title{Select Columns Using a Bitmask}
\usage{
select_by_mask(data, intMask)
}
\arguments{
\item{data}{A Numeric vector or matrix in tidy form}

\item{intMask}{An Integer vector whose length equals number of columns in data}
}
\value{
A matrix containing the columns of data for which intMask is 1
}
\description{
Selects columns from a matrix.  A column is included in the output when the
corresponding mask value is 1.
}
\examples{
e2 <- embed(example2_data, 17)
tn <- e2[ , 1]
pn <- e2[ ,2:17]
msk <- integer(16)
msk[c(1,2,3,4,6,7,9)] <- 1  # select these columns
p <- select_by_mask(pn, msk)
gamma_test(predictors = p, target = tn)

mask <- int_to_intMask(15, 16)     # pick out the first four columns
p <- select_by_mask(pn, mask)
get_Mlist(predictors = p, target = tn, by = 50)
}
