% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-package.R
\docType{package}
\name{sr-package}
\alias{sr}
\alias{sr-package}
\title{sr: Smooth Regression - The Gamma Test and Tools}
\description{
Finds causal connections in precision data, finds lags and embeddings in time series, guides and evaluates the performance of neural networks and other smooth models, gives a mathematically sound answer to the over-training problem. Smooth regression is based on the Gamma test, which measures smoothness in a multivariate relationship. 'sr' includes the Gamma test and search techniques that use it. See references in the help for function `gamma_test`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://smoothregression.com}
  \item \url{https://github.com/haythorn/sr/}
  \item Report bugs at \url{https://github.com/haythorn/sr/issues}
}

}
\author{
\strong{Maintainer}: Wayne Haythorn \email{support@smoothregression.com}

Authors:
\itemize{
  \item Antonia Jones (Principal creator of the Gamma test)
}

Other contributors:
\itemize{
  \item Sam Kemp (Wrote the original code for the Gamma test in R) [contributor]
}

}
\keyword{internal}
