% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_search.R
\name{fe_search}
\alias{fe_search}
\title{Full Embedding Search}
\usage{
fe_search(predictors, target, prog_bar = TRUE, n_neighbors = 10, eps = 0)
}
\arguments{
\item{predictors}{A vector or matrix whose columns are proposed inputs to a
predictive function}

\item{target}{A vector of double, the output variable that is to be predicted}

\item{prog_bar}{Logical, set this to FALSE if you don't want progress bar displayed}

\item{n_neighbors}{Integer number of near neighbors to use in RANN search,
passed to gamma_test}

\item{eps}{The error limit for the approximate near neighbor search.  This
will be passed to gamma_test, which will pass it on to the ANN near neighbor search.  Setting
this greater than zero can significantly reduce search time for large data sets.}
}
\value{
An invisible data frame with two columns, an integer mask representing a particular subset of
the predictors, and the value of Gamma using those predictors and the target.
This is sorted from lowest to highest Gamma.  This data.frame also has an attribute
named target_V, containing the target variance which is the same for all rows.
To get the vratio for any row, divide its Gamma by target_v.
}
\description{
Calculates gamma for all combinations of a set of input predictors
}
\details{
Given a set of predictors and a target that is to be predicted, this search
will run the gamma test on every combination of the inputs.  It returns the
results in order of increasing gamma, so the best combinations of inputs for
prediction will be at the beginning of the list.  As this is a fully
combinatoric search, it will start to get slow beyond about 16 inputs.  By default,
fe_search will display a progress bar showing the time to completion.

fe_search() returns a data.frame of two elements, a sorted vector of
Gamma values, and an integer containing the mask representing the inputs
used to calculate that Gamma, selected from the predictors passed to the search.
To reconstruct each data set, use int_to_intMask and select_by_mask as shown
in their examples.
}
\examples{
e6 <- embed(mgls, 7)
t <- e6[ ,1]
p <- e6[ ,2:7]
full_search <- fe_search(predictors = p, target = t)
}
