% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Mlist.R
\name{get_Mlist}
\alias{get_Mlist}
\title{Discover how Gamma varies with sample size}
\usage{
get_Mlist(
  predictors,
  target,
  plot = TRUE,
  caption = "",
  show = "Gamma",
  from = 20,
  to = length(target),
  by = 20
)
}
\arguments{
\item{predictors}{A Numeric vector or matrix whose columns are proposed
inputs to a predictive relationship}

\item{target}{A Numeric vector, the output variable that is to be predicted}

\item{plot}{A logical, set this to FALSE if you don't want the plot}

\item{caption}{Character string to be used as caption for the plot}

\item{show}{Character string, if it equals "vratio", vratios will be plotted,
otherwise Gamma is plotted}

\item{from}{Integer length of the first data sample, as passed to seq}

\item{to}{Integer maximum length of sample to test, passed to seq}

\item{by}{Integer increment in lengths of successive windows, passed to seq}
}
\value{
An invisible data frame containing Gamma values, vratios and the
length of the associated data set
}
\description{
Produces a list of M values (sample sizes), and the associated Gammas
}
\examples{
he <- embed(henon_x, 17)
t <- he[ , 1]
p <- he[ ,2:17]
ml <- get_Mlist(p, t)
get_Mlist(p, t, caption = "this data")
ml <- get_Mlist(p, t, plot = FALSE)
}
