% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-package.R
\docType{package}
\name{sr-package}
\alias{sr}
\alias{sr-package}
\title{sr: Smooth Regression - The Gamma Test and Tools}
\description{
Finds causal connections in precision data, finds lags and embeddings in time series, guides training of neural networks and other smooth models, evaluates their performance, gives a mathematically grounded answer to the over-training problem. Smooth regression is based on the Gamma test, which measures smoothness in a multivariate relationship. Causal relations are smooth, noise is not. 'sr' includes the Gamma test and search techniques that use it. References: Evans & Jones (2002) \doi{10.1098/rspa.2002.1010}, AJ Jones (2004) \doi{10.1007/s10287-003-0006-1}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://smoothregression.com}
  \item \url{https://github.com/haythorn/sr/}
  \item Report bugs at \url{https://github.com/haythorn/sr/issues}
}

}
\author{
\strong{Maintainer}: Wayne Haythorn \email{support@smoothregression.com}

Authors:
\itemize{
  \item Antonia Jones (Principal creator of the Gamma test)
}

Other contributors:
\itemize{
  \item Sam Kemp (Wrote the original code for the Gamma test in R) [contributor]
}

}
\keyword{internal}
